requirements.txt
opencv-python
mss
pynput
numpy
pyyaml
PyQt6
